requirements.txt
# Core MCP framework
fastmcp>=0.1.3

# Outlook / Microsoft Graph
msal>=1.30.0
requests>=2.32.3
python-dotenv>=1.0.1
pyjwt>=2.9.0

# Gmail / Google API
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.151.0

# AI/LLM
groq>=0.4.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2

# Google Sheets
gspread>=5.12.0
google-auth-httplib2>=0.1.1

# HTML Parsing
beautifulsoup4>=4.12.0

# Utilities
email-validator>=2.1.0
